Reorganized the files by move all the header files to include folder.
