+++ test/include/test.h
@@ -1,7 +1,7 @@
 #ifndef __TEST_HEADER__
 #define __TEST_HEADER__
 
-#include "../../src/cards_management.h"
+#include "../../include/cards_management.h"
 
 extern Deck_t* g_draw_pile_test;                     /* remaining cards to draw */
 extern Deck_t* g_discard_pile_test;                  /* discarded cards */
